+++ src/components/client/UserDetails/components/DetailsPhone.tsx
@@ -3,6 +3,7 @@
 import { useTranslations } from "next-intl";
 
 import Modal from "@/components/common/Modal";
+import FormButton from "@/components/common/FormButton";
 import {
   Details,
   hidePhone,
@@ -38,7 +39,7 @@ export function DetailsPhone({ id, number }: DetailsPhoneProps) {
               placeholder={t("placeholder.phone")}
               defaultValue={number}
             />
-            <Details.Submit isEditable />
+            <FormButton isEditable />
           </form>
         </Modal>
       </Details.View>
