+++ src/components/client/UserDetails/components/common/Details.tsx
@@ -6,7 +6,6 @@ import { DetailsPhone } from "@/components/client/UserDetails/components/Details
 import { DetailsView } from "@/components/client/UserDetails/components/common/DetailsView";
 import { DetailsPassword } from "@/components/client/UserDetails/components/DetailsPassword";
 import { DetailsInput } from "@/components/client/UserDetails/components/common/DetailsInput";
-import { DetailsSubmit } from "@/components/client/UserDetails/components/common/DetailsSubmit";
 
 export default function Details({ children }: { children: ReactNode }) {
   return (
@@ -21,5 +20,4 @@ Details.Name = DetailsName;
 Details.Email = DetailsEmail;
 Details.Phone = DetailsPhone;
 Details.Input = DetailsInput;
-Details.Submit = DetailsSubmit;
 Details.Password = DetailsPassword;
